plugins/tools/skills/deep-research/scripts/requirements.txt
yt-dlp>=2024.3.10
notebooklm-py[browser]>=0.3.3
beautifulsoup4>=4.12.0
lxml>=5.1.0
requests>=2.31.0
playwright>=1.42.0
readability-lxml>=0.8.1
html2text>=2024.2.26
python-dotenv>=1.0.0
